% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{al1_group_neut}
\alias{al1_group_neut}
\title{Neutralisation (demean) by group
Cross-sectionally neutralized (demeaned) by groups}
\usage{
al1_group_neut(x, groups = NULL)
}
\arguments{
\item{x}{(xts) time series matrix.}

\item{groups}{(factor) groups allocation, one for each column.}
}
\description{
Neutralisation (demean) by group
Cross-sectionally neutralized (demeaned) by groups
}
