% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{al1_param_defs}
\alias{al1_param_defs}
\title{Global function definitions.}
\usage{
al1_param_defs()
}
\arguments{
\item{x}{(xts) time series matrix.}

\item{y}{(xts) time series matrix.}
}
\description{
Global function definitions.
}
