% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{al1_rank}
\alias{al1_rank}
\title{Cross-sectional rank
Proportion of firms that have an observed value less than or equal to the
observed value.}
\usage{
al1_rank(x)
}
\arguments{
\item{x}{(xts) time series matrix.}
}
\value{
(xts)
}
\description{
Cross-sectional rank
Proportion of firms that have an observed value less than or equal to the
observed value.
}
