% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{al1_roll_rank}
\alias{al1_roll_rank}
\title{Time series rank}
\usage{
al1_roll_rank(x, k)
}
\arguments{
\item{x}{an object (representing a series of observations).}

\item{k}{integer width of the rolling window. Must be odd for \code{rollmedian}.}
}
\description{
Time series rank
}
