% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{al1_scale}
\alias{al1_scale}
\title{Rescale x such that sum(abs(x)) == a, column-wise.}
\usage{
al1_scale(x, a = 1)
}
\arguments{
\item{x}{(xts) time series matrix.}

\item{a}{(numeric) target sum(abs(.)) value of every column after scaling.}
}
\value{
(xts)
}
\description{
Rescale x such that sum(abs(x)) == a, column-wise.
}
