% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{al1_whichmax}
\alias{al1_whichmax}
\title{Time series argument max (rolling)}
\usage{
al1_whichmax(data, width)
}
\arguments{
\item{data}{the data to be used (representing a series of observations).}

\item{width}{numeric vector or list. In the simplest case this is an integer
    specifying the window width (in numbers of observations) which is aligned
    to the original sample according to the \code{align} argument. Alternatively,
    \code{width} can be a list regarded as offsets compared to the current
    time, see below for details.}
}
\description{
Time series argument max (rolling)
}
