% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphas.R
\name{alpha_common_parameters}
\alias{alpha_common_parameters}
\title{Definitions of common parameters used in alpha functions.}
\usage{
alpha_common_parameters()
}
\arguments{
\item{k}{(numeric) vector of constants used to evaluate the alpha.}

\item{op}{(xts) matrix with asset open prices.}

\item{cl}{(xts) matrix with asset close prices.}

\item{hi}{(xts) matrix with asset high prices.}

\item{lo}{(xts) matrix with asset low prices.}

\item{vol}{(xts) matrix with volume traded.}

\item{dvol}{(xts) matrix with dollar volume traded.}

\item{r}{(xts) matrix with logarithmic returns.}

\item{cap}{(xts) matrix with market capitalisation values. #FIXME:}
}
\description{
Definitions of common parameters used in alpha functions.
}
