% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alpha_run}
\alias{alpha_run}
\title{Evaluate an alpha function given data (xts) and constant arguments.
Extracts relevant data from data argument and returns xts alpha results.}
\usage{
alpha_run(constants, dat, alpha_fun)
}
\arguments{
\item{constants}{(numeric) vector of constants to be passed to alpha fun.}

\item{dat}{(named list) each entry an xts matrix (e.g. cl= closing prices,
vol=volumes).}

\item{alpha_fun}{(function) the alpha function to be evaluated.}
}
\value{
(xts matrix) alpha value computed for every period and every asset.
}
\description{
Evaluate an alpha function given data (xts) and constant arguments.
Extracts relevant data from data argument and returns xts alpha results.
}
\examples{
To evaluate alpha1 against constants (20, 5) and data contained in `d`:
alpha_run(c(20, 5), d, alpha1)
}
