% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sym_data_tidy}
\alias{sym_data_tidy}
\title{Convert historical stock price and volume data to long format.}
\usage{
sym_data_tidy(d)
}
\arguments{
\item{d}{(data frame) price and volume data from `quantmod::getSymbols` call.}
}
\description{
Convert historical stock price and volume data to long format.
}
