% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{syms_data_get}
\alias{syms_data_get}
\title{Get price and volume data on several ticker symbols in one long data frame.}
\usage{
syms_data_get(syms = test_syms, ...)
}
\arguments{
\item{syms}{(character) ticker symbol(s) to process}

\item{...}{further arguments passed to `quantmod::getSymbols`}
}
\description{
Get price and volume data on several ticker symbols in one long data frame.
}
