% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{syms_data_xts}
\alias{syms_data_xts}
\title{Convert a data frame with historical price data to a list of xts matrices.}
\usage{
syms_data_xts(d)
}
\arguments{
\item{d}{(data frame) historical price data from `syms_data_get`}
}
\value{
(named list) each entry an xts time-series matrix.
}
\description{
Convert a data frame with historical price data to a list of xts matrices.
}
